requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
openai==0.8.0
gunicorn==20.1.0